CDAP-15658 fixed tests for generic databases + added tests for schemas and fields validation
