+++ database-commons/src/main/java/io/cdap/plugin/db/batch/source/AbstractDBSource.java
@@ -29,6 +29,7 @@ import io.cdap.cdap.etl.api.Emitter;
 import io.cdap.cdap.etl.api.PipelineConfigurer;
 import io.cdap.cdap.etl.api.batch.BatchRuntimeContext;
 import io.cdap.cdap.etl.api.batch.BatchSourceContext;
+import io.cdap.cdap.etl.api.validation.InvalidConfigPropertyException;
 import io.cdap.cdap.etl.api.validation.InvalidStageException;
 import io.cdap.cdap.internal.io.SchemaTypeAdapter;
 import io.cdap.plugin.common.LineageRecorder;
@@ -84,7 +85,7 @@ public abstract class AbstractDBSource extends ReferenceBatchSource<LongWritable
     if (importQueryString.toUpperCase().contains("WHERE $CONDITIONS AND")) {
       importQueryString = importQueryString.replaceAll("(?i)" + Pattern.quote("$CONDITIONS AND"), "");
     } else if (importQueryString.toUpperCase().contains("WHERE $CONDITIONS")) {
-      importQueryString = importQueryString.replaceAll("(?i)"  + Pattern.quote("WHERE $CONDITIONS"), "");
+      importQueryString = importQueryString.replaceAll("(?i)" + Pattern.quote("WHERE $CONDITIONS"), "");
     } else if (importQueryString.toUpperCase().contains("AND $CONDITIONS")) {
       importQueryString = importQueryString.replaceAll("(?i)" + Pattern.quote("AND $CONDITIONS"), "");
     } else if (importQueryString.toUpperCase().contains("$CONDITIONS")) {
@@ -380,7 +381,30 @@ public abstract class AbstractDBSource extends ReferenceBatchSource<LongWritable
     }
 
     private void validateSchema(Schema actualSchema) {
-      DBUtils.validateSourceSchema(actualSchema, getSchema());
+      validateSchema(actualSchema, getSchema());
+    }
+
+    static void validateSchema(Schema actualSchema, Schema configSchema) {
+      if (configSchema == null) {
+        throw new InvalidConfigPropertyException("Schema should not be null or empty", SCHEMA);
+      }
+      for (Schema.Field field : configSchema.getFields()) {
+        Schema.Field actualField = actualSchema.getField(field.getName());
+        if (actualField == null) {
+          throw new InvalidConfigPropertyException(String.format("Schema field '%s' is not present in actual record",
+                                                                 field.getName()), SCHEMA);
+        }
+        Schema actualFieldSchema = actualField.getSchema().isNullable() ?
+          actualField.getSchema().getNonNullable() : actualField.getSchema();
+        Schema expectedFieldSchema = field.getSchema().isNullable() ?
+          field.getSchema().getNonNullable() : field.getSchema();
+
+        if (!actualFieldSchema.equals(expectedFieldSchema)) {
+          throw new IllegalArgumentException(
+            String.format("Schema field '%s' has type '%s' but found '%s' in input record",
+                          field.getName(), expectedFieldSchema.getType(), actualFieldSchema.getType()));
+        }
+      }
     }
 
     @Nullable
